refactor(XLRowData, XLComments, tests): ♻️ use GSL for type safety in numeric conversions
- add gsl/gsl header for type-safe narrowing casts
- replace raw size() and arithmetic with gsl::narrow and gsl::narrow_cast
- remove unused exception variable in XLComments catch block

+++ OpenXLSX/headers/XLRowData.hpp
@@ -17,6 +17,7 @@
 #include <list>
 #include <memory>
 #include <vector>
+#include <gsl/gsl>
 
 // ===== OpenXLSX Includes ===== //
 #include "OpenXLSX-Exports.hpp"
@@ -283,10 +284,10 @@ namespace OpenXLSX
             // ===== If the container value_type is XLCellValue, the values can be copied directly.
             if constexpr (std::is_same_v<typename T::value_type, XLCellValue>) {
                 // ===== First, delete the values in the first N columns.
-                deleteCellValues(values.size());    // 2024-04-30: whitespace support
+                deleteCellValues(gsl::narrow<uint16_t>(values.size()));    // 2024-04-30: whitespace support
 
                 // ===== Then, prepend new cell nodes to current row node
-                auto colNo = values.size();
+                auto colNo = gsl::narrow<uint16_t>(values.size());
                 for (auto value = values.rbegin(); value != values.rend(); ++value) {    // NOLINT
                     prependCellValue(*value, colNo);    // 2024-04-30: whitespace support: this is safe because only prependCellValue (with
                                                         // whitespace support) touches the row data
@@ -296,7 +297,7 @@ namespace OpenXLSX
 
             // ===== If the container value_type is a POD type, use the overloaded operator= on each cell.
             else {
-                auto range = XLRowDataRange(*m_rowNode, 1, values.size(), getSharedStrings());
+                auto range = XLRowDataRange(*m_rowNode, 1, gsl::narrow<uint16_t>(values.size()), getSharedStrings());
                 auto dst   = range.begin();    // 2024-04-30: whitespace support: safe because XLRowDataRange::begin invokes whitespace-safe
                                                // getCellNode for column 1
                 auto src = values.begin();
